pyproject.toml
[project]
name = "fastapi-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "imagekitio>=5.2.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
